暂存工作

+++ server/src/main/kotlin/git/imitater967/otc/accounting/bo/AccountingSubject.kt
@@ -1,5 +1,5 @@
 package git.imitater967.otc.accounting.bo
 
-class AccountingSubject(val name: String,val id: UInt) {
+class AccountingSubject(val name: String,val id: UShort) {
 
 }
